+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Restricted Input - Release Notes
 
+## UNRELEASED
+
+* Provide specific strategy for Android KitKat Chromium based webviews. Fixes issue where only 4 characters could be input
+* Provide helper method for determining if the browser supports formatting
+
 ## 1.0.11 (2017-02-22)
 
 * Correct detection of Samsung browsers to disable formatting in old versions